CRAYSAT-1237: Skip setting active version
This commit changes the default behavior so that running
cray-product-catalog-update does not set a particular version to
'active', unless the SET_ACTIVE_VERSION variable is set in the
environment.

Test Description: After building the image locally, I ran the image
to add content to the product catalog, with and without the
SET_ACTIVE_VERSION in the environment.